Increase DMA copy throughput by using deeper FIFOs
This commit changes the FIFOs in the DMA Copy Unit to be 2-element
FIFOs rather than 1-element FIFOs, so that they can be enqueued to
and dequeued from in the same cycle.
rl_mm2s_fifof_deq has been split into two rules in order to avoid
a scheduling conflict.
It also adds some print statements to allow seeing throughput when
debugging in simulation.