Add explicit conditions to force correct operation
The explicit conditions added here shouldn't have any effect on the
operation of the Copy Unit, since their behaviour should be the exact
same as the behaviour of the underling implicit conditions, but it seems
like they do have an effect, at least in simulation.
The effect is probably on the scheduling properties of the code. The new
scheduling also means that the copy unit has reduced throughput, at
least as measured in simulation copying to and receiving from the
loopback in AXI_Utils.