Add halting support and full CHERI cap checking
This commit:
Adds proper mid-transfer halting functionality so that if an
error occurs during a transfer the bus is not locked up.
Adds support for properly handling CHERI errors on
transfers and reporting these in the registers.
Fixes a bug where capabilities fetched from SG were being put in the
internal register module in reverse word order.
Avoids writing buffer descriptor capabilities into memory, as this would
remove the tags.