Copy Unit: only reset after hold external reset
The ethernet block was producing a flit of data on the metadata stream
on the first cycle where the reset was triggered, and this flit was not
being cleared when the shims' clear method was being called because they
were occurring in the same cycle.
This led to the copy unit being in the "META_RECEIVE_S2MM" state between
packets rather than "IDLE" so when a new packet arrived the first flit
would not get dropped.